hoist Error type definition into new module
