+++ src/Language/Haskell/Liquid/GHC/SpanStack.hs
@@ -1,6 +1,6 @@
 {-# LANGUAGE BangPatterns #-}
 
-module Language.Haskell.Liquid.Types.SpanStack
+module Language.Haskell.Liquid.GHC.SpanStack
    ( -- * Stack of positions
      Span (..)
    , SpanStack
@@ -11,19 +11,20 @@ module Language.Haskell.Liquid.Types.SpanStack
      -- * Using Stacks
    , srcSpan
 
+     -- * Creating general spans
+   , showSpan
    ) where
 
-import SrcLoc
+import           SrcLoc
 import qualified Var
-import CoreSyn                          hiding (Tick, Var)
-import Name                             (getSrcSpan)
-import FastString                       (fsLit)
-import Language.Haskell.Liquid.GHC.Misc (tickSrcSpan, showPpr)
+import           CoreSyn                hiding (Tick, Var)
+import           Name                   (getSrcSpan)
+import           FastString             (fsLit)
 import Data.Maybe                       (listToMaybe, fromMaybe)
-import Language.Haskell.Liquid.Misc     (firstJust)
+import Language.Haskell.Liquid.GHC.Misc (tickSrcSpan, showPpr)
 
 -- | Opaque type for a stack of spans
-newtype SpanStack = SpanStack {unStack :: [(Span, SrcSpan)]}
+newtype SpanStack = SpanStack { unStack :: [(Span, SrcSpan)] }
 
 --------------------------------------------------------------------------------
 empty :: SpanStack
@@ -67,5 +68,7 @@ maybeSpan d sp
   | isGoodSrcSpan sp = Just sp
   | otherwise        = d
 
+--------------------------------------------------------------------------------
 showSpan :: (Show a) => a -> SrcSpan
+--------------------------------------------------------------------------------
 showSpan = mkGeneralSrcSpan . fsLit . show
