+++ src/Language/Haskell/Liquid/Transforms/ANF.hs
@@ -39,11 +39,10 @@ import           Language.Fixpoint.Types            (anfPrefix)
 import           Language.Haskell.Liquid.Misc       (concatMapM)
 import           Language.Haskell.Liquid.GHC.Misc   (MGIModGuts(..), showPpr, symbolFastString)
 import           Language.Haskell.Liquid.Transforms.Rec
-import           Language.Haskell.Liquid.UX.Errors
+import           Language.Haskell.Liquid.Types.Errors
 import qualified Language.Haskell.Liquid.Types.SpanStack as Sp
 import           Data.Maybe                       (fromMaybe)
 import           Data.List                        (sortBy, (\\))
--- import           Control.Applicative
 
 
 --------------------------------------------------------------------------------
